update show form add container connect
