+++ src/components/showForm/component/utils.js
@@ -1,5 +1,5 @@
 export const KeyCodes = {
-  comma: 188,
-  enter: 13,
+    comma: 188,
+    enter: 13,
 }
 export const delimiters = [KeyCodes.comma, KeyCodes.enter]
